requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Azure SDKs
openai==1.30.5                     # Azure OpenAI client
azure-search-documents==11.6.0b4   # Azure AI Search
azure-storage-blob==12.20.0        # Azure Blob Storage
azure-identity==1.16.0             # Managed identity / DefaultAzureCredential

# Document parsing
pymupdf==1.24.3                    # PDF → text (fitz)
python-docx==1.1.2                 # .docx support (optional)

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
tenacity==8.3.0                    # Retry logic for API calls
tiktoken==0.7.0                    # Token counting
rich==13.7.1                       # Pretty CLI output
